Support concurrent writers in fake_chronicle_kv
Rebalance unit tests have been intermittently failing as a rebalance
makes concurrent writes to chronicle_kv via multiple processes. The
existing fake_chronicle_kv could not support this. A transaction
could take a snapshot, and apply it's updates to it. During this time
another transaction could take a snapshot, apply updates, and store the
result. The first transaction, upon completion, would lose any writes
stored by the second.

Update fake_chronicle_kv to perform compare and swap operations on the
snapshot state, and to retry in the event of a failure.

Change-Id: I82454a206939c9a9bad4e76839e9ae3a1ba5cf55
Reviewed-on: https://review.couchbase.org/c/ns_server/+/231616
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>