MB-67857: Only catch menelaus_cbauth_worker noproc errors
If an individual call to menelaus_cbauth_worker:sync/1 fails, such as
the call timed out, we should propogate that crash toc
b_creds_rotation, so that we don't continue rotating the passwords
when a cbauth instance might not have the new passwords.

However, this code had a catch for a reason, which was to handle the
race between getting the list of pids to call and a worker going down.
To retain handling for this race, we need to explicitly catch noproc
exits, while allowing all other exceptions to be propagated to the
caller.

Change-Id: I244bfc12422ccbee549d640566402934f8c8115c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/231713
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker