MB-67886: Manually fetch pip for cluster_tests
Now that ensurepip isn't bundled with the python we ship couchbase
with, cluster_tests can't use it to install requirements.

Initially we fixed this with a temporary solution to make use of the
python already installed on the Jenkins workers:
https://review.couchbase.org/c/ns_server/+/231838

This approach doesn't work on every machine, since not every machine
has python with support for venv and pip.

Instead we have to manually download pip, and just use that.

Change-Id: I7ae64dcb734537ca720be25911c2b6d7c89c56bb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/231854
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>