MB-66812: Use sha-256 hash instead of full VC for metakv rev
This change addresses the issue where vector clock revisions were too
long. Switch to the new format:
- Use a SHA256 hash of the vector clock

This does not change anything about the metakv vector clocks internally.
Services use `rev` to perform a CAS operation. `rev` is queried on a
node and passed on subsequent mutate operations - to ensure that the
underlying value is updated atomically. During a mutate operation, `rev`
is computed afresh and compared (strict equality) with `rev` in the
incoming request. If they differ, the operation is aborted.

This does not require any changes in the API nor to services because:
- Services treat `rev` as an opaque identifier. `rev` is queried using a
  GET. The raw bytes are sent in subsequent mutate operations. (PUT
  sends rev as raw bytes in a form URL-encoded body, DELETE as percent-
  encoded raw bytes in the query string.)
- Services use `rev` on a node-local basis. There is no need to
  consider distinct `rev` formats on nodes running different versions.
- During an upgrade, a service is restarted along with ns_server/cbauth.
  i.e. a node running a particular version always uses the same format
  for metakv Get/Set/Delete.

Filed MB-68058 for further cleanup.

Change-Id: I9f9bf165ac486fe213f1219a1a5b52e78e8b3d98
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232169
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Neelima Premsankar <neelima.premsankar@couchbase.com>