MB-67857: Fix flakiness of menelaus_cbauth cbauth_sync_t
The test can race with setting up menelaus_cbauth_worker, because the
setup doesn't wait for this to be set up. The other tests weren't
impacted because they implicitly waited for this by waiting for the
initial UpdateDB call to be made. To ensure that cbauth_sync_t, and
any future tests, are not flaky, the setup function should wait for the
initial notification.

Change-Id: If8688e8d720196b49260fc2f2f65701d5e3eef77
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232520
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Restriction Checker