MB-61292: Change proof calculation for encr at rest keys
Current algorithm for calculating proof for a key uses this
very key for encryption. This may cause issues when key is
can't be used for encryption (e.g. there is no connection
to AWS). Inability to calculate proof causes inability
to save that key on disk.
This leads to several usability problems and confusions, such as:
1. constant save retries causes pollution of logs with errors even
   when the key is not used.
2. integrity tokens can't be rotated until all keys can be saved
   successfully.
3. when the key is modified and can not be saved, the previous
   version of this key will be used for encryption of other keys,
   which may lead to confusion (e.g. AWS credentials file is
   changed, but the system keeps using prev creds file).

This change's goal is to only use those parts of the key that
are available locally for proof calculation.

Change-Id: Ie1cdab90a81545a3a650910e58c658e16b17e00c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232626
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>