MB-68255 - UI to allow "force full encryption"
The idea of "partial encryption" is new to the morpheus release - it
occurs when either:

- encryption is turned on, but pre-existing data is still unencrypted

- encryption is turned off, but some existing data is still encrypted

This "partial encryption" state was previously only indicated by a
tiny orange icon, with a tooltip lacking any explanation. Plus, there
was nothing the user could do to change the situation.

This change adds text as well as an icon to indicate whether
encryption is full or partial. It also adds more explanation to the
tooltip about why the situation came about. Finally, it provides a
button to invoke "force encryption" which either completely encrypts
or decrypts the data in question.

Alas the components in question existed in both AngularJS and Angular
8 code, so I needed to implement the solution twice. (I tried
downgrading the newer component, but even with AI help it wouldn't
work.)

Change-Id: I646915ee7ddcf01eda880198e20261b6ef397553
Reviewed-on: https://review.couchbase.org/c/ns_server/+/233056
Well-Formed: Restriction Checker
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>