MB-68456: [BP] Make full eviction default bucket magma policy
If magma is not specified as the storage backend, then the default
policy will remain value eviction. This happens because is_magma/4
defaults to assuming that the bucket is couchstore, which is no longer
the case.

Update the function to default to magma, and correct the tests that
break accordingly.

This was inadvertently merged to master, rather than morpheus.

Change-Id: I08ceef683e6e08d87575052d59073eb8704685ea
Reviewed-on: https://review.couchbase.org/c/ns_server/+/234386
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>