MB-68545: Set explicit verify_none for addNode, completeJoin
Erlang 26 introduced an incompatible change:

* the default value for the verify option is now verify_peer
  instead of verify_none.

In trinity, we fixed this for log collection in
https://review.couchbase.org/c/ns_server/+/208889

We didn't need to fix this for /addNode and /completeJoin because we
fixed it inadvertently in
https://review.couchbase.org/c/ns_server/+/173592

Now that we are upgrading to Erlang 26 in neo, we need to backport that
fix. This change essentially backports 405a76e, but only the
refactoring logic needed to add verify_none when appropriate for
addNode and completeJoin.

Change-Id: I7b0a0ad1483c628762ae6eb6a37156e199aef391
Reviewed-on: https://review.couchbase.org/c/ns_server/+/233918
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>