MB-68878: Add back {resource_management, enabled} flag
In provisioned profile, the flag was removed in 8.0.0, because it would
be true for all profiles, so it was considered redundant to include.
However, it actually isn't redundant, because in mixed mode clusters,
the check still needs to be performed, to avoid allowing enabling
guardrails in mixed mode self-managed clusters. Since the check will
still be performed in mixed mode clusters, the flag needs to be
retained.

Change-Id: I88afb9c679ffd4ae8580702083583d5f76325306
Reviewed-on: https://review.couchbase.org/c/ns_server/+/234907
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker