MB-68870 Pass addr family requirements to cont backup
If the cluster was configured for IPv6 only ns_server was not passing
this information to continuous backup. However, even if ns_server had
done so, continuous backup doesn't support the settings.

So this change is dependent on continuous backup supporting the
settings.

Change-Id: Ie9d99daf2a8ade2895bf9d3e14e0b5ca4a7850df
Reviewed-on: https://review.couchbase.org/c/ns_server/+/234924
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>