MB-69632: Use 'backup' compat_version when restoring users
Use the compat_version info in the results of a 'backup' to determine
whether or not to make backwards compatible role changes. This is to
handle the case, for example:

* the backup results have a user with the "ro_admin" role. If the backup
  was taken prior to 8.0 then a restore should change the user to have
  "ro_admin" + "ro_security_admin".

  If the backup was taken on 8.0 then the restore should not change the
  user. Without the compat_version we can't differentiate the two cases.

The compat_version is being added in the 8.0.1 release with a 8.0
compat_version. So a backup taken on 8.0 will have user role changes
when restored. But this is the same issue that happens prior to this
change.

Moving forward, any restore of a backup result should check the
compat_version and only upgrade users based on that info.

Restores of compat versions from the future will result in an error.

Change-Id: Ifb7a07c7a58a6f184ab97a205532ec0a41dc96c8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/237227
Well-Formed: Restriction Checker
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>